fix: make all 9 non-resistant patterns compiler-resistant at -O3
Rewrote generators for AL-3, DS-2, DS-3, HR-4, IS-2, IS-5, MI-2,
MI-3, SR-1 to survive -O3 optimization using noinline helpers in
separate translation units (helper.c).

Key techniques per pattern:
- AL-3: noinline comparison helper + adversarial all-ones data
- DS-2: noinline alloc/free helpers with volatile pointers
- DS-3: noinline pass-by-value helper with 256-byte struct
- HR-4: noinline check helper with volatile guards
- IS-2: noinline expensive transform for outlier path
- IS-5: two noinline kernels (restrict vs non-restrict) in helper.c
- MI-2: noinline zero_buffer helper in separate TU
- MI-3: noinline alloc/free helpers with volatile pointers
- SR-1: rewritten from Form B (algebraic) to Form A (noinline
  function hoist) — same approach as SR-4

Also added proper test harnesses (SLOW_CODE_HERE markers) for
AL-1, DS-4, IS-1, MI-4, SR-2 which were previously skipped by
the batch tester.

Faithfulness: 555/555 (100%).

+++ dataset/DS_2/DS-2_v006/fast.c
@@ -1,9 +1,10 @@
-void fast_ds2_v006(float *results,float *input,int n,int chunk){
-    float *tmp=(float*)malloc(chunk*sizeof(float));
+#include <stdlib.h>
+void fast_ds2_v006(double *results,double *input,int n,int chunk){
+    double *tmp=(double*)malloc(chunk*sizeof(double));
     for(int i=0;i<n;i+=chunk){
         int sz=(i+chunk<=n)?chunk:(n-i);
         for(int j=0;j<sz;j++) tmp[j]=input[i+j]*input[i+j];
-        float sum=0; for(int j=0;j<sz;j++) sum+=tmp[j];
+        double sum=0; for(int j=0;j<sz;j++) sum+=tmp[j];
         results[i/chunk]=sum;
     }
     free(tmp);
