+++ dataset/IS_1/IS-1_v008/fast.c
@@ -1,7 +1,7 @@
-double fast_is1_v008(double *A, double *B, int n) {
-    double sum = 0.0;
+float fast_is1_v008(float *A, float *B, int n) {
+    float sum = 0.0f;
     for (int i = 0; i < n; i++) {
-        if (A[i] == 0.0 || B[i] == 0.0) continue;
+        if (A[i] == 0.0f || B[i] == 0.0f) continue;
         sum += A[i] * B[i];
     }
     return sum;
