+++ dataset/IS_1/IS-1_v008/slow.c
@@ -1,5 +1,5 @@
-double slow_is1_v008(double *A, double *B, int n) {
-    double sum = 0.0;
+float slow_is1_v008(float *A, float *B, int n) {
+    float sum = 0.0f;
     for (int i = 0; i < n; i++) {
         sum += A[i] * B[i];
     }
