+++ dataset/MI_2/MI-2_v014/fast.c
@@ -1,3 +1,3 @@
-void fast_mi2_v014(double *out,double *A,double *B,int n){
-    for(int i=0;i<n;i++) out[i]=A[i]*B[i]+(double)1.0;
+void fast_mi2_v014(float *out,float *A,float *B,int n){
+    for(int i=0;i<n;i++) out[i]=A[i]*B[i]+(float)1.0f;
 }
