+++ dataset/MI_4/MI-4_v008/fast.c
@@ -1,4 +1,4 @@
-void fast_mi4_v008(double *out, double *A, double *B, int rows, int cols) {
+void fast_mi4_v008(float *out, float *A, float *B, int rows, int cols) {
     for (int i = 0; i < rows; i++) {
         for (int j = 0; j < cols; j++) {
             out[i * cols + j] = A[i * cols + j] + B[i * cols + j];
