+++ dataset/MI_4/MI-4_v008/slow.c
@@ -1,4 +1,4 @@
-void slow_mi4_v008(double *out, double *A, double *B, int rows, int cols) {
+void slow_mi4_v008(float *out, float *A, float *B, int rows, int cols) {
     for (int j = 0; j < cols; j++) {
         for (int i = 0; i < rows; i++) {
             out[i * cols + j] = A[i * cols + j] + B[i * cols + j];
